api/requirements.txt
bcrypt==3.1.4
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
passlib==1.7.1
psycopg2==2.7.6.1
PyJWT==1.6.4
